requirements.txt
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
playwright>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
